I finished the project (at least I think I did...)

+++ client/src/components/Trips/Trips.tsx
@@ -1,6 +1,7 @@
 import { useContext, useState, useEffect } from "react";
 import { UsePageContext } from "../ContextPage/ContextPage";
 import { TripIDContext } from '../ContextDetailsId/ContextDetailsId';
+import {TokenContext} from '../ContextToken/ContextToken';
 import './Trips.css'
 
 
@@ -18,12 +19,15 @@ interface Trip {
   
 
 export default function Trips() {
+  const contextToken = useContext(TokenContext);
+  if(!contextToken) return null;
   const context = useContext(UsePageContext);
   if (!context) return null;
   const contextID = useContext(TripIDContext)
   if(!contextID) return null;
   const { setCurrentTripID } = contextID;
   const { setCurrentPage } = context;
+  const { token } = contextToken
   const [trips, setTrips] = useState<Trip[]>([]);
   
   useEffect(() => {
@@ -34,10 +38,14 @@ export default function Trips() {
   }, []);
 
   const handleDelete = (tripId: string) => {
+    if(token === null) { 
+      alert("you are not a registered user")
+      return
+    }
     fetch(`http://localhost:3000/api/trips/${tripId}`, {
         method: 'DELETE',
         headers: {
-          Authorization: 'test-token'
+          Authorization: token
         }
       })
       .then(() => {
@@ -51,7 +59,13 @@ export default function Trips() {
     setCurrentPage("TripDetails");
     setCurrentTripID(tripId);
   };
-
+  const hendleManagerStuff = () => {
+    if(token === null) { 
+      alert("you are not a registered user")
+      return
+    }
+    setCurrentPage('NewTripForm')
+  }
 
 
   return (
@@ -59,14 +73,14 @@ export default function Trips() {
       <h1>Trips</h1>
       <div className="pageMovButtons">
         <button onClick={() => setCurrentPage('Home')}>Home</button>
-        <button onClick={() => setCurrentPage('NewTripForm')}>NewTripForm</button>
+        <button onClick={hendleManagerStuff }>NewTripForm</button>
         
       </div>
       <div>
         {trips.map((trip) => (
           <div key={trip.id} className="trip-card" >
             <h3>{trip.name}</h3>
-            <div> {trip.price} $</div>
+            
             <img src={trip.image} alt={trip.name} />
             <button onClick={() => handleDelete(trip.id)}>delete</button>
             <button onClick={() => handleShowDetails(trip.id)}>Tripdata</button>
